work/requirements.txt
pytube3
moviepy
pytesseract
elasticsearch>=7.0.0,<8.0.0
opencv-python
pymongo